just cast it to an unsigned long long G(CC/++) (or the stdlib it's using) seems to define a uint64_t as an unsigned long, whereas clang(++) (or the stdlib that macOS is using) defines it as an unsigned long long. whatever

+++ xpdbg_software/src/xpDBG_window.cc
@@ -143,7 +143,7 @@ xpDBG_window::xpDBG_window(int		argc,
 			asprintf(&disassembly_text,
 					 "%s0x%016llx:\t%s\t\t%s\n",
 					 disassembly_text,
-					 insn[i].address,
+					 (unsigned long long)insn[i].address,
 					 insn[i].mnemonic,
 					 insn[i].op_str);
 		}
